refactor: move feature importance to its own helper module
